Fix compilation issues due to cudnn 12.2 TK.

+++ samples/test_list.cpp
@@ -2067,7 +2067,7 @@ TEST_CASE("Conv two global scales", "[frontend][fusion][conv global scale]") {
     for (size_t i = 0; i < (size_t)YSize; i++) {
         half afterConvOutput = afterConv.hostPtr[i];
         half finalOutput = Y.hostPtr[i];
-        half globalScaleOutput = afterConvOutput * scale1.hostPtr[0] * scale2.hostPtr[0];
+        half globalScaleOutput = (float)afterConvOutput * scale1.hostPtr[0] * scale2.hostPtr[0];
         float diff         = getError(finalOutput, globalScaleOutput);
         if (diff < 0) diff = -diff;
         if (diff > THRESHOLD) { numErrors++;}
